+++ src/pane.cc
@@ -535,5 +535,5 @@ Fsi_plugin_arg ()
       (FARPROC)xpiSetPanePos,
       0,
     };
-  return make_fixnum (long (procs));
+  return make_fixnum (long long (procs));
 }
